// src/types/enemies.ts
export type EnemyType =
  | 'AntsInMyEyesJohnson'
  | 'Concerto'
  | 'FloopyDoops'
  | 'Froopy5'
  | 'Froopy14'
  | 'FroopyTommy'
  | 'GeneralGromflamite'
  | 'Gobo'
  | 'GromflamiteSoldier'
  | 'Hamurai'
  | 'Hemorrhage'
  | 'Jaguar'
  | 'Kiara'
  | 'MrNimbus'
  | 'NoobNoob'
  | 'PrincessPoneta'
  | 'RisottoGroupon'
  | 'ScaryTerry'
  | 'Schleemypants'
  | 'Shnyuk'
  | 'ShonoopyBloopers'
  | 'Squanchy'
  | 'TargetBot'
  | 'Terryfold'